[이용기] ThemeParkControlloer 매핑
